// components/layouts/Y2KCollageBackground/types.ts
export interface GlitterStar {
  id: number;
  x: number;
  y: number;
  size: number;
  delay: number;
  duration: number;
}

export interface FloatingHeart {
  id: number;
  x: number;
  y: number;
  duration: number;
  delay: number;
}

export interface Sticker {
  id: number;
  x: number;
  y: number;
  emoji: string;
  duration: number;
  delay: number;
  rotation: number;
}